NF: Move sync files to their own subfolders
