+++ AnkiDroid/src/main/res/layout/sync_progress_layout.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="wrap_content"
     android:layout_height="wrap_content"
-    tools:context=".SyncActionProvider">
+    tools:context=".sync.SyncActionProvider">
 
     <androidx.appcompat.widget.AppCompatImageButton
         android:id="@+id/button"
